Refactor to get rid of global allocator singleton
