+++ libafl_frida/src/lib.rs
@@ -7,13 +7,18 @@ It can report coverage and, on supported architecutres, even reports memory acce
 pub mod asan_rt;
 /// The `LibAFL` frida helper
 pub mod helper;
+/// The frida-asan allocator
+pub mod alloc;
+/// Handling of ASAN errors
+pub mod asan_errors;
+
 // for parsing asan cores
 use libafl::bolts::os::parse_core_bind_arg;
 // for getting current core_id
 use core_affinity::get_core_ids;
 
 /// A representation of the various Frida options
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, serde::Serialize, serde::Deserialize)]
 #[allow(clippy::struct_excessive_bools)]
 pub struct FridaOptions {
     enable_asan: bool,
